+++ eggs/Naaya/CHANGELOG.rst
@@ -1,5 +1,6 @@
 4.0.32 (unreleased)
 -------------------
+* Editor tool: add email tab in insert link plugin [dumitval]
 
 4.0.31 (2016-01-25)
 -------------------